fix(app): rename rune files to .svelte.ts for SSR compilation
Files using Svelte 5 runes ($state, $derived) must have .svelte.ts
extensions for the compiler to process them. Plain .ts files pass
through without rune compilation, causing ReferenceError: $state is
not defined during SSR (500 on beehive dashboard).
