+++ app/src/lib/client/sse.svelte.ts
