+++ app/src/lib/stores/environment.svelte.ts
