+++ app/src/lib/stores/index.ts
@@ -1,3 +1,3 @@
-export { environment } from "./environment";
-export { runnersStore } from "./runners";
-export { notificationsStore } from "./notifications";
+export { environment } from "./environment.svelte";
+export { runnersStore } from "./runners.svelte";
+export { notificationsStore } from "./notifications.svelte";
